requirements.txt
feedparser
requests
python-dotenv
thefuzz[speedup]